changed Projects.js to Projects.jsx and made changes requested from pr
